type of field was changed

+++ src/main/java/uz/yeoju/yeoju_app/payload/resDto/module/GetGradesOfStudent.java
@@ -4,7 +4,7 @@ import java.sql.Timestamp;
 
 public interface GetGradesOfStudent {
     String getId();
-    Integer getGrade();
+    Float getGrade();
     Timestamp getTime();
     Timestamp getCreatedAt();
     String getDescription();
